Debug de front y back

+++ restful-web-services/src/main/java/com/in28minutes/rest/webservices/restfulwebservices/todo/TodoHardcodedService.java
@@ -24,7 +24,7 @@ public class TodoHardcodedService {
 	
 	public Todo save(Todo todo) {
 		if(todo.getId()==-1 || todo.getId()==0) {
-			todo.setId((long) ++idCounter);
+			todo.setId(++idCounter);
 			todos.add(todo);
 		} else {
 			deleteById(todo.getId());
